requirements.txt
aiomysql==0.0.22
anyio==3.5.0
asgiref==3.5.0
CacheControl==0.12.10
cachy==0.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
cleo==0.8.1
click==8.0.4
clikit==0.6.2
crashtest==0.3.1
databases==0.5.5
distlib==0.3.4
fastapi==0.75.0
filelock==3.6.0
greenlet==1.1.2
h11==0.13.0
html5lib==1.1
httptools==0.4.0
idna==3.3
importlib-metadata==4.11.3
keyring==23.5.0
lockfile==0.12.2
msgpack==1.0.3
mysqlclient==2.1.0
packaging==20.9
pastel==0.2.1
pexpect==4.8.0
pkginfo==1.8.2
platformdirs==2.5.1
poetry==1.1.13
poetry-core==1.0.8
ptyprocess==0.7.0
pydantic==1.9.0
pylev==1.4.0
PyMySQL==0.9.3
pyparsing==3.0.7
python-dotenv==0.19.2
PyYAML==6.0
requests==2.27.1
requests-toolbelt==0.9.1
shellingham==1.4.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.32
starlette==0.17.1
tomlkit==0.10.0
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
virtualenv==20.13.4
watchgod==0.8
webencodings==0.5.1
websockets==10.2
zipp==3.7.0
